chore: cargo fmt
Forgot cargo fmt was built in so this is the first time it was used. Hopefully cleaned up more than what it broke - I may have to go redo the tabbing in the gui controls and see if I can disable cargo from detabbing my tabbed lines. Makes it easier to group the GUI in my head with the tabs.